firmware_ECCharging: If battery full, wait for drain.
Instead of exiting when the battery is full, drain the battery to a
chargable level, and then resume the test.

To discharge the battery, the test will request the EC to go to
discharge mode.  Looking at all EC config files, there appear to be a
handful of boards that have batteries, but don't support the discharge
mode, but those boards don't exist in the test lab.

BUG=b:152530996
TEST=Ran locally with eldrid & kindred boards.

Change-Id: I1f7f486d4d4749f63c6a8638b8606ff0258f3c46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2687204
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Brent Peterson <brentpeterson@chromium.org>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>