[Autotest] Make is_up_fast() faster for for higher counts
Currently, count will increase at 1 second per n-1 (so a count of 3
takes 2 seconds, atleast). Instead, lets loop it to exit on the first
found. This function is designed to simply say "is up" not, "is up for
all counts". This can be seen on the original return just checking the #
of replies is > 0

BUG=None
TEST=dummy_pass (where this method is called)

Change-Id: Ib5e5026e7ad8f5ef7348f47d0092bc095b6a289b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2694046
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>