Changed storage_qual_v2_NAME to hardware_StorageQualV2.NAME
Tauto tests normally follow the convention that the test name is the
same as the directory name.  This test was breaking that convention,
which was also breaking the Lead/TL/Bug mapping between the Green
Stainless Tauto spreadsheet (http://go/cros-green-stainless-tauto) and
the Green Stainless Dashboard (http://go/cros-automated-test-dashboard).

BUG=b:163057136
TEST=test_that -b ${BOARD} ${DUT} ${NEW_TEST_NAME}

Confirmed that the test name is recognized as expected bit tests were
not run to completion.

Change-Id: I2538dbdf62dc8d3bc89267fbb17bef86d27c8600
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2692362
Reviewed-by: Ross Zwisler <zwisler@google.com>
Commit-Queue: Matt Davis <mattedavis@google.com>
Tested-by: Matt Davis <mattedavis@google.com>