Update test prefix with moblab_RunSuite_
Tauto tests normally follow the convention that the test name is the
same as the directory name.  This test was breaking that convention,
which was also breaking the Lead/TL/Bug mapping between the Green
Stainless Tauto spreadsheet (http://go/cros-green-stainless-tauto) and
the Green Stainless Dashboard (http://go/cros-automated-test-dashboard).

BUG=b:163057136
TEST=test_that -b ${BOARD} ${DUT} ${NEW_TEST_NAME}

Confirmed that new autotest uses the new name but did not run
the test itself.

Change-Id: Ie81b208d9593884350ff7774899fbf5455995bab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2353387
Reviewed-by: Ross Zwisler <zwisler@google.com>
Commit-Queue: Matt Davis <mattedavis@google.com>
Tested-by: Matt Davis <mattedavis@google.com>