Added test_name prefix to BootDevice and StorageQual tests.
Tauto tests normally follow the convention that the test name is the
same as the directory name.  This test was breaking that convention,
which was also breaking the Lead/TL/Bug mapping between the Green
Stainless Tauto spreadsheet (http://go/cros-green-stainless-tauto) and
the Green Stainless Dashboard (http://go/cros-automated-test-dashboard).

BUG=b:163057136
TEST=test_that -b ${BOARD} ${DUT} ${NEW_TEST_NAME}

Confirmed that the test name is recognized as expected. Tests were
not run to completion.

Change-Id: I9f9dac48dca8281b55d61229fbdc2b5eda8f317c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2698616
Reviewed-by: Ross Zwisler <zwisler@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Matt Davis <mattedavis@google.com>
Tested-by: Matt Davis <mattedavis@google.com>