cellular: Add flag to enable containments.
To avoid failures that prevent the rest of the test to be executed when
a known problem exist and the solution is in development, the flag
enable_temp_containments is introduced to allow specific containments
to be added for different boards. By default, the flag is enabled.

BUG=b:180002232
TEST=Execute cellular_Smoke.verizon on jinlon.

Change-Id: Ia473203e972543faaed09c2fdd017ea6453d522e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2702433
Tested-by: Dinesh Kumar Sunkara <dsunkara@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Commit-Queue: Harpreet Grewal <harpreet@chromium.org>