autotest: Fix flipped throughput limits
Throughput limits were recently added to the WiFi perf suite. There are
separate limits for the tx and rx WiFi modes, and the limits for the
modes were mistakenly flipped with one another. This CL applies the
correct limits for the correct modes. It also changes the order of the
limit definitions to align with the order in which the tests were
defined in previously to help avoid confusion.

BUG=b:177004163
TEST=Run various perf tests and check results.

Change-Id: I4cca8870a46882c72654ce8c57283851736037aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2705926
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>