firmware_ECLidSwitch: Use power states.
Instead of using the mode_aware_switcher to watch the suspend resume,
trust that the EC is reporting the correct power state. This saves about
1 minute in the test, and should make it less flaky as well. The risk is
that we don't have all the expected power states in place, and there
will be new failures if are other shutdown states besides S5 and G3.

Also the "shutdown and wake immediately" test wasn't very immediate,
since it was waiting several seconds between attempts to check the power
state.

This was written by Brent Peterson, I just got it working and tested it.

BUG=b:179100599
TEST=5x local: kindred, 2x skylab: careena, treeya360, akemi, dragonair
ignoring failures that happen in test setup/teardown.

Change-Id: Id4726f6068a6ec38534af1bde79d968682023285
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2740057
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>