Add a check for timezone jump due to log format change.
http://crrev.com/c/2703984 attempted to fix autoupdate_EndToEnd test but
failed due to possibly timezone info not being set on the test server
side. This hack checks if the time passed between autoupdate and reboot
is greater than 7 hours and if so assumes that the time jump is due to
the log format change. The reboot is supposed to take no longer than
900s so this assumption should be reasonable.

BUG=chromium:1178930
TEST=None

Change-Id: Ia2174490bc90767c1af6eefb688ee5ea01f5f655
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2711666
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Yuta Hijikata <ythjkt@chromium.org>