firmware_ECCharging: Fix test failures.
Fixed 2 problems.

"Unhandled AutoservRunError: command execution error": Sometimes the
stressapptest doesn't have enough free memory to run, as it is
incidental to the test and it doesn't really matter if it succeeds or
not, just ignore the status code.

"The device is not charging. Is the test run with AC plugged?" On some
devices, it appears that the battery takes a little time to start
charging after being switched out of discharge mode, so wait a few
seconds before failing the test after discharging for things to
stabilize. I have verified this fix on a krane device in skylab.

BUG=b:152530996
TEST=Tried 3 machines in skylab, was able to verify one of the 2 issues is resolved.

Change-Id: I053f727a073a3ac90a70a43270ef914acea25d95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2714630
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>