network_WiFi_UpdateRouter: update Gale test AP version
For various reasons (see b/177380545 and b/180617663) we need
to deploy a new image to the Gale test APs in the near future.

Unfortunately it's difficult to build a new R85-based image
(same issue as b/172414187#comment17 cropping up again). The
simplest and most future-proof solution is another version-
bump to a release which has the required fixes merged already,
so the new images are built from the R88 release branch.

In addition to that, there are three or four CLs we would like
to patch in (see the motivating bugs above):

  crrev.com/c/2689214: Modifies watchdog behavior on panic
  crrev.com/c/2689215: Fixes k->vfree() in essedma driver
  crrev.com/c/2689216: Drops rmmod support in essedma driver
  crrev.com/c/2689499: Enable CONFIG_MBO in hostapd

The watchdog fix is still up in the air, and I started testing
with an image lacking the CONFIG_MBO change, so the following
configurations of image have been tested:

  (A) R88-13597.90.0-b4793198:  Watchdog and both essedma fixes
  (B) R88-13597.100.0-b4802789: Watchdog, essedma, and CONFIG_MBO
  (C) R88-13597.100.0-b4802787: essedma and CONFIG_MBO only

The first image was tested thoroughly against various types
of DUT, and the latter two were spot-checked just in case.
Full test details:

  R88 (A) + Zork:    https://paste.googleplex.com/6073576076607488
  R88 (A) + Hatch:   https://paste.googleplex.com/4951859602128896
  R88 (A) + Octopus: https://paste.googleplex.com/6515605654994944
  R88 (A) + Atlas:   https://paste.googleplex.com/4880736453656576
  R88 (B) + Hatch:   https://paste.googleplex.com/5135483580448768
  R88 (C) + Hatch:   https://paste.googleplex.com/6213652274216960

In addition to running WiFi test suites, I manually verified that
patching in CL 2689216 stops check_ethernet.hook from causing a
panic, and that CL 2689214 appears to fix reboot-on-panic.

This CL currently modifies network_WiFi_UpdateRouter to use
option B (all the CLs), but I don't really have an opinion
on whether we ought to include the watchdog change. If the
consensus is to go with option C instead I'll change it.

BUG=b:177380545, b:180617663
TEST=See Above

Change-Id: I8fb63bc5cb0c3f49e40917310addcb58a8153554
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2713727
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Will Donnelly <wgd@google.com>
Commit-Queue: Will Donnelly <wgd@google.com>