Bluetooth: Search BT logs for ncmd=0 occurrences on failure
We discovered that one of our unsupported chipsets will return success
on a command but with ncmd=0, indicating that the host should send more
commands. This is then never updated, which essentially freezes the host
and prevents recovery. This change should catch such an occurrence and
surface it to stainless so we can understand the severity of this issue.

Failures are detected by locating a case of "cmd_cnt 0", indicating that
the controller is telling the host to wait before submitting more
commands, and a "cmd queued" value of 5-99. The reason for the range is
that sometimes several commands are submitted together, and it's
possible that several items are added to the queue at once.

BUG=b:181051732
TEST=LEHealth with variations of the log we're catching to confirm regex

Change-Id: Ib3e68b8db2fd382d0f22dd95cd060311edc31afd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2720388
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>