[Autotest] expose timeout option for longer tests
Currently, if a test exceeds an hour it will get warnings and turn
orange (or purple) in stainless, even if it passes. This is because
there is a hard default timeout set in ssh_host, that tests cannot set.
This CL will add in a param to the control files to allow tests to
specifically set/exceed the 1 hour limit

To use this param add: "EXTENDED_TIMEOUT = n" to the control file, where
n is the number of seconds the test should be allowed to take.

BUG=chromium:1179350
TEST=tests which timeout with/without the CL, and verify with the CL it
passes.

Change-Id: I16cdc2944d1706891aabf7e9bbca3f10a9e8935a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2728358
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>