firmware_ECCharging: Handle battery discharging.
Sometimes the battery is discharging, even when ac is attached and the
battery is not full. If the battery is in that state when the test
starts, discharge the battery a little bit to get it into an charging
state instead.

Sometimes the battery gets full again before we have a chance to start
monitoring the charging, so discharge the battery for some extra time.

Refactored the code a bit to make it more readable.

This should fix errors like:
TestFail: The device is not charging. Is the test run with AC plugged?

BUG=b:152530996
TEST=skylab: dumo, druwl, bobba

Change-Id: Idb1218bdd2c0a6d971342b0acdae5ad5129dc8ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2729695
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>