autotest: Remove network_WiFi_GTK from WiFi suites.
The Tast counterpart, tast.wifi.GTK, was promoted to stable and
is in wifi_matfunc, so remove it from wifi_matfunc.

This CL also removes it from wificell-cq, while leaving the Tast
counterpart *not* in wificell-cq. The Tast counterpart appears to
be more reliable, but still short of the 99% bar for wificell-cq.

The pass rate in past 28 days:
  wifi.GTK:         4139/(4139+87)  = 97.94% pass rate
  network_WiFi_GTK: 4760/(4760+207) = 95.83% pass rate

BUG=b:159092883
TEST=none

Change-Id: Id17ee25c1d7e7de8af90aa58e7c6c07e3a0051d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2733957
Tested-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>