autotest: Remove MultiAuth/Reassociate from WiFi suites
The tests `network_WiFi_MultiAuth` and `network_WiFi_Reassociate`
have their Tast counterparts in wifi_matfunc, so this CL removes
the originals.

This CL also removes them from wificell-cq, but their Tast ports
are *not* in wificell-cq. In both cases the Tast port appears to
be more reliable, but still short of the 99% bar for wificell-cq.

Currently the test pass rates are:

    wifi.MultiAuth:         4165/(4165+89)  = 97.9% pass rate
    network_WiFi_MultiAuth: 4697/(4697+219) = 95.5% pass rate

    wifi.Reassociate:         4165/(4165+59)  = 98.6% pass rate
    network_WiFi_Reassociate: 4707/(4707+201) = 95.9% pass rate

BUG=b:159092883
TEST=None

Change-Id: I4e357c20915ad96b7647f238d89ffe2fb217830b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2743721
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>