autotest: Ensure that modem is actually disabled.
On Qualcomm modems, disable command sends a QMI message
to put the modem in a low power state. ModemManager returns
success if the QMI message was sent successfully.
The modem takes around 1 second to actually transition to a low power state,
if an enable command is issued during this transition then we end up in a failed state.
Added check to ensure the modem actually transitioned to a low power
state before we exit the disable function.

BUG=b:181090169
TEST=test_that lazor --fast  cellular_SafetyDance.att --iterations 6

Change-Id: Id3a6332d7aac6adb96d3f44f5215f77ccfd626e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2743437
Reviewed-by: Madhav . <madhavadas@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Madhav . <madhavadas@google.com>
Commit-Queue: Madhav . <madhavadas@google.com>