Bluetooth: Split reset_peer for btpeer and for emulated device
Before the test starts and btpeers are being allocated, the reset
functionality is useful to return to a fresh state. However, sometimes
it is useful to be able to also reset the state of an emulated device in
the middle of a test, and for that the existing calls are not
sufficient.

This change allows a device to be reset from either the btpeer layer, or
from an already allocated emulated device.

BUG=b:181051299
TEST=QuickHealth on Hatch

Change-Id: I0ea00458e0e0a4fe8f9fee6d6df98bcbde7babe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2746620
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>