Bluetooth: Allow HID reboot test to function with OOBE auto-reconnect
On some platforms (ChromeBase, ChromeBox, ChromeBit), the DUT will
automatically reconnect to a paired HID device during boot. This
disagrees with our HID reboot test, which expects to reperfor the
reconnection itself. The difficulty here is that the peer device does
not automatically return to a connectable state like some peripherals
might, and so the subsequent HID test after boot will fail.

This change adds logic to reset the BT peer to a clean state during the
reboot test to ensure that it can accept the connection from the DUT
during OOBE mode. This is only done if the platform is a type that will
be automatically re-connecting.

This change also has the benefit of giving us some test coverage of this
OOBE re-connect mechanism, which right now is not tested at the platform
layer.

BUG=b:181051299
TEST=cl_HID_reports_reboot_test on ChromeBase

Change-Id: I1a580ff4a97706f51acd868192f4bf3703c23301
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2746621
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>