firmware_ECUsbPorts: force the unicode string to string
The value reads from the config is a unicode string, refer
to b/181106943#comment61 for the detail error log.
That error will make the firmware_ECUsbPorts be skipped
during the test because the string won't be matched with
unicode which starting from u'. In order to fix it,
we convert the unicode string to string.

BUG=b:181106943, b:183478411
TEST=test_that -b {BOARD} {IP} firmware_ECUsbPorts

Change-Id: Iedb216b51c6508e6681ae029b35d95645bacd86d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2755537
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>