[Autotest] Enable "fast" by default
I would like to try experimenting turning the param "Fast" on by
default. There is a lot of excess logging that happens in autotest,
especially on Green runs. This would strip out a lot of it (mostly all
of the sysinfo stuff). On fails, sysinfo would still get logged. On a
local run this drops ~25 seconds off a server test, and a few seconds
off a client.

BUG=None
TEST=cq, dummy_Pass, stub_ServerToClientPass

Change-Id: If2566d6f346d89d8e9ba47a788066f8e74b8c3ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2763205
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>