bluetooth: SetRemoteAddress in test_disconnection_by_device
The remote_address of a device may not be set when the method
test_disconnection_by_device is called.

This patch sets the remote address explicitly as a temporary
solution. A complete solution would be to set the remote
address when the corresponding PropertiesChanged signal is received.

BUG=b:182864322
TEST=The following test sequence should work.
  self.test_pairing()
  self.test_disconnection_by_device(device)

Change-Id: I968648500170507dd61dacab5d1f1018aa9242b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2780824
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Manish Mandlik <mmandlik@chromium.org>