firmware_Fingerprint: Add test case for cros_config
We're seeing that trying to get the cros_config info periodically fails.
It's unclear whether the failure is from the autotest SSH connection or
getting cros config data triggers something on the device. This test
case is meant to be a simplified version of the code that seems to
trigger the problem so that we can get a better reproduction case.

BUG=b:160271883
TEST=test_that --board hatch <IP> firmware_Fingerprint.CrosConfig

Change-Id: I5fce2ef0bf25bb927cb2621527878fc71011d438
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2768778
Reviewed-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>