Bluetooth: introduce AVL test version control
This patch brings the script and the updated control file to automated
a version control process for AVL tests to sync the test environments
among vendors, partners, and developers. The script will fetch a commit
hash of the "released" (or "stabled") version of the autotest directory
from the cloud and checkout locally.

If the process failed to fetch the commit hash from the cloud, it would
use the ToT instead. On successful sync, the test will fail with an
informative log ask the user to rerun the test. Only if the local
commit is the same as the cloud's, the test will continue to execute.

BUG=b:183067141
TEST=See b:183067141 for details

Change-Id: Iaa59abb441458283de6f7234b3915226c5b01083
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2771444
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Michael Sun <michaelfsun@google.com>
Tested-by: Michael Sun <michaelfsun@google.com>
Auto-Submit: Michael Sun <michaelfsun@google.com>