FAFT/firmware_CsmeFwUpdate: run test if RW blobs differ
Currently, it is possible for two images that report the same RW
version to actually be different from each other.  For this reason,
we will compare the downgrade blob against the two me_rw blobs in the
spi flash (ME_RW_A and ME_RW_B) and then run this test if any of the
RW blobs differ.

BUG=b:182312849
TEST=install CPFE OS 13853.0.0 and CPFE firmware 13672.109.0 on a
voxel, execute 'test_that --board=$BOARD
--autotest_dir=/mnt/host/source/src/third_party/autotest/files
$DUT_IP firmware_CsmeFwUpdate' to run the test,
and verify that the test notes the RW versions are the same but
that the blobs are different and runs the test.
TEST=install CPFE OS 13731.0.0 and CPFE firmware 13672.23.0 on a
voxel, run the test and verify that the test notes the RW versions
and contents of the RW blobs are the same and skips the test as N/A.
TEST=install CPFE OS 13731.0.0 and CPFE firmware 13672.23.0 into
RW_SECTION_A and CPFE firmware 13672.138.0 into RW_SECTION_B on a
voxel, run the test and verify that the test notes that the downgrade
blob and the RW_SECTION_B blobs differ and that it proceeds to run
the test.

Change-Id: I962c3bc3ebd36e7fdcac052c33e62170fa1848a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2773948
Tested-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>