Bluetooth: Surface BT failures in BT/Wifi coex tests
In my original implementation of co-ex tests, the test name was left
blank to run the whole batch. I didn't realize that when a batch is run,
the final test results are only returned at the batch-level, causing
test-level failures to not be surfaced to stainless. This change
provides the new test name to ensure that the test failure reasons are
provided to stainless.

BUG=b:182797631
TEST=Ran all co-ex tests with deliberate BT failure to validate

Change-Id: Ie1a6e086c5b056f87b8e8e44b7ad1d40487755ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2776154
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Michael Sun <michaelfsun@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>