FAFT: Make check_keys() check key press order
In check_keys(), when we try to remove the repeated key presses, we
sort the received key presses so that the order is unfortunately lost.

As a result, although check_keys() tests if the expected keys are
pressed, check_keys() does not test if the expected keys are pressed in
the order we want. (As of now, check_keys() implicitly requires
expected_sequence to be sorted.)

Let's fix it:
When processing the captured key presses, we can remove the repeated
key presses without lossing the order of the key presses.

BUG=b:176258249
TEST=confirm check_keys() works with unsorted expected_sequence

Change-Id: I09e6868d24e829340d1a903f5e5eef8ddf7acab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2776159
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Philip Chen <philipchen@chromium.org>