[Autotest] work around keyerror for results
This is a workaround but the error is not fatal and the
main purpose was to avoid having an error in the provisioning logs that
confuses people debugging tests.
Sometimes, for a TBD reason, the results do not have the utils_lib.DIRS
yet.

BUG=b:181691434
TEST=dummy_Pass/cq

Change-Id: Idd8f418d8d1e45be3ad909ab770dd425de4cfc90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2779813
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>