Bluetooth: Remove device after verifying RSSI
Before each bluetooth test, we discover each peer and verify adequate
RSSI. After discovery, bluetoothd puts this device into a "temporary"
state and starts a 30 second timer. At the end of the timer, the device
is removed. I've discovered that in some tests, this 30 second timeout
unfortunately can occur right when the device is meant to be pairing,
causing a test failure.

This test clears the device after discovery in the RSSI check to ensure
a fresh start to the test.

BUG=b:169633231
TEST=Several BT tests to verify the device isn't removed unexpectedly

Change-Id: Iae640cd61712606980677a5cd3aca7d90a888d91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2780561
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>