bluetooth: add expect_fail for tests that are expected to fail
This patch adds the expect_fail method to support running
a test which is expected to fail.

As an example, a test is expected to fail if the behavior is
disallowed by policy. In this case, the failure is considered
a pass.

In the following example, if a self.test_xxx(device), which is
a test with @test_retry_and_log wrapper, is expected to fail,
the usage looks like

    self.expect_fail(self.test_xxx, device)

As a result, the log message shows that the test conceptually passed.
    [*** passed: test_xxx (expected fail, actually failed)]

A little helper, expect_test, is also supported to convert

    if expected_result:
        self.test_xxx(device)
    else:
        self.expect_fail(self.test_xxx, device)

to a simple one liner

    self.expect_test(expected_result, self.test_xxx, device)

BUG=b:183471556
TEST=None

Change-Id: I3d8c37a84632022fbcea0eddcc4c7e2021598ba0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2784293
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>