Reduce network disconnection period in AU OOBE interruption test
Reduce the period during which the network is disconnected from 2
minutes to 25 seconds in autoupdate_ForcedOOBEUpdate.interrupt.network.
The ServerAliveInterval for autotest's ssh connection was reduced to 30
seconds in CL:2517561, so client tests that disconnect the DUT's
internet for greater than 30 seconds will be prematurely ended by the
server. This happens 100% on moblab, but only periodically in the lab
as the new ServerAliveInterval does not seem to affect all ssh
connections.

BUG=chromium:1172867
TEST=test_that chromeos6-row14-rack22-host1.cros autoupdate_ForcedOOBEUpdate.interrupt.network.full -b octopus -i 'octopus-release/R91-13885.0.0' --no-retry --autotest_dir ~/trunk/src/third_party/autotest/files --args="job_repo_url=http://100.115.168.195:8082/static/octopus-release/R91-13885.0.0/autotest/packages"
TEST=au_fsi on moblab with the new time patched in

Change-Id: I42742a42b7ebb3a87b785eea432f1aea75a649c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2798762
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>