bluetooth: populate _expected_result early in quick_test_test_decorator
If there is a TESTNA exception in the early stage of a quick test,
it might lead to

    AttributeError: 'bluetooth_AdapterEPHealth' object has no attribute
                    '_expected_result'

We should populate the _expected_result attribute earlier in the
quick_test_test_decorator so that the attribute will still get
populated even if there are any early exceptions.

BUG=b:184296655
TEST=Run a abluetooth_AdapterEPHealth test, e.g., on a nightfury.
Add the following test code to the test wrapper with
    skip_models=['nightfury']
The test should pass with this patch and should fail on AttributeError
without this patch.

Change-Id: Id537c685f49b16bd175ac1512557d360d82a60bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2807196
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Alain Michaud <alainm@chromium.org>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Miao-chen Chou <mcchou@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>