bluetooth: Increase resume time for peer wake
On the flaky dashboard, several devices (older ones) are reporting
resume times of 21-24s and are failing the peer wake tests. However,
this just shows the controllers are slow at picking up the wake event,
not that the software is misconfigured.

Another commit is adding the wake time as a perf measurement so increase
the allowed resume time for peer wake so these older chipsets also pass
the test if they wake.

BUG=b:184552011
TEST=Ran bluetooth_AdapterSRHealth.sr_peer_wake_le_hid

Change-Id: I57ef6ded6a2f079f00c1a14d34adfc64742be198
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2806403
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>