firmware_ECChargingState: Use battery params
In addition to asking the kernel for the battery state, check the
battery status and params from the EC. This will let us pass when the
battery is full and discharging, which is a state the kernel doesn't
track.

BUG=b:176992437
TEST=hatch/kindred, and repro'd the FULL DCHG state that caused the
failure before.
Also tested on volteer/eldrid for good measure.

Change-Id: I009df80be8ef27d45dd012aebd794c06c3285bc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2805264
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>