bluetooth: do not run bluetooth_AdapterEPHealth batch in the lab
Currently, bluetooth_AdapterEPHealth is run as a batch in the lab.
It is sufficient to run its individual quick tests. Hence, this
patch removes the batch from the running list and saves the lab
resources.

BUG=b:183471556
TEST=Observe that the bluetooth_AdapterEPHealth batch is not
running in the lab.

Change-Id: I4707ecfff6145f47307cc25b57e2f0438323b300
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2804130
Reviewed-by: Alain Michaud <alainm@chromium.org>
Commit-Queue: Miao-chen Chou <mcchou@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>