logging_UserCrash: Simplify, remove ported test.
Remove the root_crasher test, and remove logic that examines the logs.
This log-examining logic is prone to flakiness and spurious failures,
and anyway is redundant with tast.

This leaves the chronos_crasher test because that verifies that the
in_progress_integration_test variable is set appropriately.

BUG=chromium:1190212
TEST=test_that $DUT logging_UserCrash

Change-Id: I24bd8145aa6a3c3aa402c25f81ea47ce4b1f0510
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2809205
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>