bluetooth: Set connectable property when advertising
Starting from BlueZ 5.54, BlueZ requires the flag
"limited discoverable" or "general discoverable" of advertisement
data to be set, otherwise the advertisement would be ignored. As we
don't set that flag on the DUT, this would cause failures in some LE
tests if the peer is running Bluez 5.54 or later versions.

Set the connectable property of advertisement so BlueZ could set
"general discoverable" flag.

BUG=b:184701938
TEST=Pass bluetooth_AdapterLEHealth with peer running Bluez v5.54

Change-Id: Idc1ea83c62c24a007c6ccdffb6f17fb834e97584
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2809681
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Tested-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>