faft: Corrupt firmware using idempotent functions
RPCProxy.__call_faft_client() assumes the RPC function is idempotent.
That is, it assumes a second call to the function would not change the
state of the FAFT client. However, methods such as
BiosServicer.corrupt_sig() are not idempotent, because this method will
increment the value of a byte in flashrom *on each call*. This
potentially leads to an increment of 2 for the byte, and then causes
BiosServicer.restore_sig() to fail to correctly restore the byte.

This CL replaces the following BiosServicer methods:

- corrupt_{sig,body}
- restore_{sig,body}

with idempotent functions

- get_{sig,body}_one_byte
- modify_{sig,body}

so that firmware signature/body can be correctly restored. In addition,
for firmware_CorruptRecoveryCache, add BiosServicer.corrupt_mrc_cache(),
which corrupts the whole RECOVERY_MRC_CACHE section.

BUG=b:183076875
TEST=firmware_CorruptBothFwBodyAB passed on Hayato
TEST=firmware_FAFTRPC.bios passed
TEST=firmware_FAFTRPC.ec passed

Change-Id: I2600eb50d9d969615aa6ef602f682b2a2f239561
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2823240
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>