Bluetooth: Accept advertising command complete for ext adv check
To validate that all platforms are using the new extended advertising
MGMT commands, we check for extended command use in the btmon traffic.
However, the current check can sometimes be squished by btmon at it
appears in the request header, causing the test to fail to detect it.
This patch slightly changes the text that is searched so that the
request's command complete event can also be used as a passing criteria
for the check.

BUG=b:185600864
TEST=AdvHealth.adv_multiple_advertising_test on hatch

Change-Id: I6598453f28c665ef2adaed5367570ca2d0235e73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2832252
Reviewed-by: Michael Sun <michaelfsun@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>