firmware_ECUsbPorts: Fail if 0 USB A ports are found
firmware_ECUsbPorts historically probed the EC via gpioget for the USB A
port rails by guessing a name. If this method turned up empty, the test
shrugged, assumed there were no USB A ports, and returned TEST_NA.
Everyone tends to ignore TEST_NA results, until (usually at FSI time)
someone finally notices the device does have type A ports and the test
should be reporting PASS, not N/A. At this point it's a fire drill,
often resulting in firmware changes to get this fixed.

To try and flag this misconfiguration earlier in the development cycle,
fail the test if the probe turns up empty handed. A new
fw-testing-config, usb_a_port_count, can be used to explicitly bless
devices with 0 type A ports, or even specify a default expected number
of type A ports.

This CL changes the logic for determining the USB A port count as such:
1) If an explicit count is specified, use that.
2) Otherwise, count the entries in custom_usb_enable_pins, if non-zero,
   and use that.
3) Finally, fall back to probing the old way.

This may cause old boards that had previously passed FAFT (either
correctly or incorrectly) to start failing if fw-testing-config is
unspecified, or specified incompatibly. The best fix would be to
specify the correct number for "usb_a_port_count". Or, to escape these
new rules and simply go back to the old probing method, add
"usb_a_port_count": -1. This -1 value serves as both an escape
hatch, and as a fallback in case the number of USB A ports can't be
exactly specified per-model (if it varies by SKU for example).

BUG=b:186461762
TEST=run firmware_ECUsbPorts with a number of different configs

Cq-Depend: chromium:2857478
Change-Id: I3f2629419b0a31afeac4c417e04fc6142f46a75e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2859641
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>