lxc: Add support for different base_container_path
This enables keeping the base container where it currently is, but
relocating new containers to a different path.

In particular, new containers could be moved to a path that is not
overlayfs and thus allows using overlayfs for snapshotting in lxc-copy.

See also bug for context.

BUG=b:187225142
TEST=None

Change-Id: I25d7d6ce86e888399f0c931a619a11a7fa071c89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2875543
Commit-Queue: Allen Li <ayatane@chromium.org>
Commit-Queue: Congbin Guo <guocb@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Tested-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>