Restore stateful in autoupdate_Basic.m2n.full.
Previously when this test was added to the CQ (and release builders) it
caused DUTs to have different versions of rootfs and stateful and caused
problems in the lab with tests running the wrong version of tests for
the current build.

This adds a stateful restore to the end of the test.

BUG=chromium:1202617
TEST=test_that <DUT> autoupdate_Basic.m2n.full

Change-Id: I6d7e7807f2c03818a5c446ff891cfedc1e0c8da5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2887000
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>